refactor: rename the structure
