+++ cnab/banks/banco_brasil/CNAB240/remessa/registro3Q.py
@@ -1,4 +1,4 @@
-from cnab.base.cnab_240 import CNAB240Registro3
+from cnab.base.remessa.CNAB240 import CNAB240Registro3
 from cnab.core.field import CNABFieldInteger, CNABFieldAlfa
 
 class BancoBrasil240Registro3Q(CNAB240Registro3):
