+++ cnab/banks/caixa/CNAB240/remessa/registro0.py
@@ -1,4 +1,4 @@
-from cnab.base.cnab_240 import CNAB240Registro0
+from cnab.base.remessa.CNAB240 import CNAB240Registro0
 from cnab.core.field import (
     CNABFieldInteger,
     CNABFieldAlfa,
