+++ cnab/banks/itau/CNAB400/remessa/registro0.py
@@ -1,4 +1,4 @@
-from cnab.base.cnab_400 import CNAB400Registro0
+from cnab.base.remessa.CNAB400 import CNAB400Registro0
 from cnab.core.field import (
     CNABCreatedDateField,
     CNABFieldInteger,
