+++ cnab/banks/itau/CNAB400/remessa/registro1.py
@@ -1,8 +1,8 @@
 from typing import Optional
-from cnab.base.cnab_400 import CNAB400Registro1
+from cnab.base.remessa.CNAB400 import CNAB400Registro1
 from cnab.core.field import CNABFieldInteger, CNABFieldAlfa, CNABFieldDecimal, CNABFieldDate
 from cnab.core.especie import CNABFieldEspecieTitulo
-from cnab.base.registro import Registro
+from cnab.base.remessa import RegistroRemessa
 from .registro2 import ItauCnab400Registro2
 
 
@@ -54,7 +54,7 @@ class ItauCnab400Registro1(CNAB400Registro1):
     filler6 = CNABFieldAlfa("", length=1, default=" ", required=True)
     numero_registro = CNABFieldInteger("", length=6, default="0", required=True)
 
-    def __init__(self, header: Optional[Registro], parent: Optional[Registro], **kwargs: dict):
+    def __init__(self, header: Optional[RegistroRemessa], parent: Optional[RegistroRemessa], **kwargs: dict):
         super().__init__(header, parent, **kwargs)
         self.inserir_multa(**kwargs)
