+++ cnab/banks/santander/CNAB240/remessa/registro1.py
@@ -1,4 +1,4 @@
-from cnab.base.cnab_240 import CNAB240Registro1
+from cnab.base.remessa.CNAB240 import CNAB240Registro1
 from cnab.core.field import (
     CNABFieldInteger,
     CNABFieldAlfa,
