+++ cnab/base/remessa/CNAB240/registro0.py
@@ -1,5 +1,4 @@
-from cnab.base.registro import Registro
-from cnab.base.registro_remessa import RegistroRemessa
+from cnab.base.remessa import RegistroRemessa
 from .registro1 import CNAB240Registro1
 
 class CNAB240Registro0(RegistroRemessa):
@@ -12,7 +11,7 @@ class CNAB240Registro0(RegistroRemessa):
         self.counter = 0
         super().__init__(None, None, **kwargs)
 
-    def append(self, child: Registro):
+    def append(self, child: RegistroRemessa):
         super().append(child)
         self.counter += 1
