+++ cnab/base/remessa/CNAB240/registro5.py
@@ -1,14 +1,13 @@
 from typing import Optional
-from cnab.base.registro import Registro
-from cnab.base.registro_remessa import RegistroRemessa
+from cnab.base.remessa import RegistroRemessa
 from .registro1 import CNAB240Registro1
 
 class CNAB240Registro5(RegistroRemessa):
     lote: CNAB240Registro1 = None
 
     def __init__(self,
-        header: Optional["Registro"],
-        parent: Optional["Registro"],
+        header: Optional[RegistroRemessa],
+        parent: Optional[RegistroRemessa],
         lote: CNAB240Registro1,
         **kwargs: dict,
     ):
