+++ cnab/base/remessa/CNAB400/registro0.py
@@ -1,12 +1,11 @@
-from cnab.base.registro import Registro
-from cnab.base.registro_remessa import RegistroRemessa
+from cnab.base.remessa import RegistroRemessa
 
 class CNAB400Registro0(RegistroRemessa):
     def __init__(self, **kwargs: dict):
         self.counter = 1
         super().__init__(None, None, **kwargs)
 
-    def append(self, child: Registro):
+    def append(self, child: RegistroRemessa):
         super().append(child)
         self.counter += 1
