+++ cnab/base/remessa/CNAB400/registro1.py
@@ -1,11 +1,9 @@
 from typing import Optional
-from cnab.base.registro import Registro
-from cnab.base.registro_remessa import RegistroRemessa
+from cnab.base.remessa import RegistroRemessa
 
-class CNAB400Registro2(RegistroRemessa):
-    def __init__(self, header: Optional[Registro], parent: Optional[Registro], **kwargs: dict):
+class CNAB400Registro1(RegistroRemessa):
+    def __init__(self, header: Optional[RegistroRemessa], parent: Optional[RegistroRemessa], **kwargs: dict):
         super().__init__(header, parent, **kwargs)
-        
         self.init_numero_registro()
 
     def init_numero_registro(self):
@@ -14,3 +12,7 @@ class CNAB400Registro2(RegistroRemessa):
         if hasattr(self, 'numero_sequencial'):
             self.numero_sequencial = self.header.counter
 
+    def append(self, child: RegistroRemessa):
+        super().append(child)
+
+        self.header.counter += 1
