+++ cnab/base/remessa/__init__.py
@@ -1,0 +1,4 @@
+from .registro_remessa import RegistroRemessa
+from .remessa import Remessa
+
+__all__ = ['RegistroRemessa', 'Remessa']
